// src/types/lotes/AlterarLote.ts
export type LotePut = {
  numeroIdentificador: string | null;
  nomeEntregador: string | null;
  nomeRecebedor: string | null;
  valorEstimado: number | string | null;
  valorHoraEstimado: number | string | null;
  dataEntrada: string | null;
  dataPrevistaSaida: string | null;
  dataInicio: string | null;
  dataSaida: string | null;
};